Move rest of router from client dir, into router package
