+++ client/modules/router/helpers.js
@@ -1,10 +1,4 @@
-import { BlazeLayout } from "meteor/kadira:blaze-layout";
 import Router from "./main";
-
-//
-// Layout container uses body
-//
-BlazeLayout.setRoot("body");
 
 //
 // pathFor
