+++ imports/plugins/core/router/client/index.js
@@ -1,0 +1,1 @@
+import "./startup";
