+++ package.json
@@ -41,6 +41,7 @@
     "font-awesome": "^4.7.0",
     "griddle-react": "0.7.1",
     "handlebars": "^4.0.6",
+    "history": "^4.6.1",
     "i18next": "7.1.3",
     "i18next-browser-languagedetector": "^1.0.1",
     "i18next-localstorage-cache": "^1.1.0",
@@ -80,7 +81,8 @@
     "react-measure": "^1.4.6",
     "react-nouislider": "^2.0.0",
     "react-onclickoutside": "^5.10.0",
-    "react-router-dom": "^4.0.0",
+    "react-router": "^4.1.1",
+    "react-router-dom": "^4.1.1",
     "react-select": "^1.0.0-rc.3",
     "react-simple-di": "^1.2.0",
     "react-taco-table": "^0.5.0",
